Tests/LibWeb/Text/expected/HTML/HTMLDetailsElement-coalesce-events.txt

details9.open set to true
details9.open is still true
target.open is false
this should be called
